# openapi-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/docs":{"get":{"operationId":"DocsController_getDocs","parameters":[],"responses":{"200":{"description":""}},"tags":["Docs"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"200":{"description":"Login successful"},"400":{"description":"Invalid credentials"}},"summary":"User login","tags":["Auth"]}},"/auth/google":{"get":{"operationId":"AuthController_googleAuth","parameters":[],"responses":{"200":{"description":""}},"summary":"Initiate Google OAuth flow","tags":["Auth"]}},"/auth/google/redirect":{"get":{"operationId":"AuthController_googleAuthRedirect","parameters":[],"responses":{"200":{"description":"Google login successful"}},"summary":"Google OAuth callback","tags":["Auth"]}},"/users/register":{"post":{"operationId":"UsersController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":""}},"tags":["Users"]}},"/users/all":{"get":{"operationId":"UsersController_getUsers","parameters":[{"name":"search","required":true,"in":"query","schema":{"type":"string"}},{"name":"bloodType","required":true,"in":"query","schema":{"type":"string"}},{"name":"userType","required":true,"in":"query","schema":{"type":"string"}},{"name":"minAge","required":true,"in":"query","schema":{"type":"number"}},{"name":"maxAge","required":true,"in":"query","schema":{"type":"number"}},{"name":"country","required":true,"in":"query","schema":{"type":"string"}},{"name":"startDate","required":true,"in":"query","schema":{"type":"string"}},{"name":"endDate","required":true,"in":"query","schema":{"type":"string"}},{"name":"page","required":true,"in":"query","schema":{"type":"number"}},{"name":"limit","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/me":{"get":{"operationId":"UsersController_getMe","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/{id}":{"get":{"operationId":"UsersController_getUserById","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/users/update":{"put":{"operationId":"UsersController_updateUser","parameters":[],"responses":{"200":{"description":""}},"tags":["Users"]}},"/blood-type/create":{"post":{"operationId":"BloodTypeController_createBlood","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BloodGroupDto"}}}},"responses":{"201":{"description":""}},"tags":["BloodType"]}},"/role/create":{"post":{"operationId":"RoleController_createRole","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRoleDto"}}}},"responses":{"201":{"description":""}},"tags":["Role"]}},"/role/get/{id}":{"get":{"operationId":"RoleController_getRole","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Role"]}},"/role/all":{"get":{"operationId":"RoleController_getAllRoles","parameters":[{"name":"hospitalId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Role"]}},"/role/update/{id}":{"patch":{"operationId":"RoleController_updateRole","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateRoleDto"}}}},"responses":{"200":{"description":""}},"tags":["Role"]}},"/role/delete/{id}":{"delete":{"operationId":"RoleController_deleteRole","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Role"]}},"/role/assign/{id}":{"put":{"operationId":"RoleController_assignPermissions","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Role"]}},"/role/remove/{id}":{"patch":{"operationId":"RoleController_removePermissions","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Role"]}},"/permission/create":{"post":{"operationId":"PermissionController_createPermission","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePermissionDto"}}}},"responses":{"201":{"description":""}},"tags":["Permission"]}},"/permission/get/{id}":{"get":{"operationId":"PermissionController_getPermission","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Permission"]}},"/permission/all":{"get":{"operationId":"PermissionController_getAllPermissions","parameters":[],"responses":{"200":{"description":""}},"tags":["Permission"]}},"/permission/update/{id}":{"patch":{"operationId":"PermissionController_updatePermission","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePermissionDto"}}}},"responses":{"200":{"description":""}},"tags":["Permission"]}},"/permission/delete/{id}":{"delete":{"operationId":"PermissionController_deletePermission","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Permission"]}},"/blood-bank/create":{"post":{"operationId":"BloodBankController_createBloodBank","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBloodBankDto"}}}},"responses":{"201":{"description":""}},"tags":["BloodBank"]}},"/blood-bank/get/{id}":{"get":{"operationId":"BloodBankController_getBloodBank","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["BloodBank"]}},"/blood-bank/all":{"get":{"operationId":"BloodBankController_getAllBloodBanks","parameters":[{"name":"hospitalId","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["BloodBank"]}},"/blood-bank/update/{id}":{"patch":{"operationId":"BloodBankController_updateBloodBank","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateBloodBankDto"}}}},"responses":{"200":{"description":""}},"tags":["BloodBank"]}},"/blood-bank/delete/{id}":{"delete":{"operationId":"BloodBankController_deleteBloodBank","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["BloodBank"]}},"/blood-bank/by-hospital/{hospitalId}":{"get":{"operationId":"BloodBankController_getBloodBanksByHospital","parameters":[{"name":"hospitalId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["BloodBank"]}}},"info":{"title":"BloodDonor API","description":"API documentation for BloodDonor Server","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"LoginUserDto":{"type":"object","properties":{}},"CreateUserDto":{"type":"object","properties":{}},"BloodGroupDto":{"type":"object","properties":{}},"CreateRoleDto":{"type":"object","properties":{}},"UpdateRoleDto":{"type":"object","properties":{}},"CreatePermissionDto":{"type":"object","properties":{}},"UpdatePermissionDto":{"type":"object","properties":{}},"CreateBloodBankDto":{"type":"object","properties":{}},"UpdateBloodBankDto":{"type":"object","properties":{}}}}}